+++ sdk/ai/Azure.AI.Inference/src/Customized/EmbeddingsClient.cs
@@ -15,6 +15,8 @@ namespace Azure.AI.Inference
     //   Modified code for the ChatCompletionsClient
 
     /// <summary> The Embeddings service client. </summary>
+    [CodeGenSuppress("Embed", typeof(EmbeddingsOptions), typeof(ExtraParameters?), typeof(CancellationToken))]
+    [CodeGenSuppress("EmbedAsync", typeof(EmbeddingsOptions), typeof(ExtraParameters?), typeof(CancellationToken))]
     public partial class EmbeddingsClient
     {
         /// <summary> Initializes a new instance of EmbeddingsClient. </summary>
@@ -42,17 +44,20 @@ namespace Azure.AI.Inference
         /// The method makes a REST API call to the `/embeddings` route on the given endpoint.
         /// </summary>
         /// <param name="embeddingsOptions"></param>
-        /// <param name="extraParams">
-        /// Controls what happens if extra parameters, undefined by the REST API,
-        /// are passed in the JSON request payload.
-        /// This sets the HTTP request header `extra-parameters`.
-        /// </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <exception cref="ArgumentNullException"> <paramref name="embeddingsOptions"/> is null. </exception>
         /// <include file="../Generated/Docs/EmbeddingsClient.xml" path="doc/members/member[@name='EmbedAsync(EmbeddingsOptions,ExtraParameters?,CancellationToken)']/*" />
-        public virtual async Task<Response<EmbeddingsResult>> EmbedAsync(EmbeddingsOptions embeddingsOptions, ExtraParameters? extraParams = null, CancellationToken cancellationToken = default)
+        public virtual async Task<Response<EmbeddingsResult>> EmbedAsync(EmbeddingsOptions embeddingsOptions, CancellationToken cancellationToken = default)
         {
             Argument.AssertNotNull(embeddingsOptions, nameof(embeddingsOptions));
+            ExtraParameters? extraParams = null;
+
+            // CUSTOM CODE NOTE:
+            //   If AdditionalProperties are provided, the decision has been made to default extraParams to "PassThrough"
+            if (embeddingsOptions.AdditionalProperties != null && embeddingsOptions.AdditionalProperties.Count > 0)
+            {
+                extraParams ??= ExtraParameters.PassThrough;
+            }
 
             using RequestContent content = embeddingsOptions.ToRequestContent();
             RequestContext context = FromCancellationToken(cancellationToken);
@@ -65,22 +70,113 @@ namespace Azure.AI.Inference
         /// The method makes a REST API call to the `/embeddings` route on the given endpoint.
         /// </summary>
         /// <param name="embeddingsOptions"></param>
-        /// <param name="extraParams">
-        /// Controls what happens if extra parameters, undefined by the REST API,
-        /// are passed in the JSON request payload.
-        /// This sets the HTTP request header `extra-parameters`.
-        /// </param>
         /// <param name="cancellationToken"> The cancellation token to use. </param>
         /// <exception cref="ArgumentNullException"> <paramref name="embeddingsOptions"/> is null. </exception>
         /// <include file="../Generated/Docs/EmbeddingsClient.xml" path="doc/members/member[@name='Embed(EmbeddingsOptions,ExtraParameters?,CancellationToken)']/*" />
-        public virtual Response<EmbeddingsResult> Embed(EmbeddingsOptions embeddingsOptions, ExtraParameters? extraParams = null, CancellationToken cancellationToken = default)
+        public virtual Response<EmbeddingsResult> Embed(EmbeddingsOptions embeddingsOptions, CancellationToken cancellationToken = default)
         {
             Argument.AssertNotNull(embeddingsOptions, nameof(embeddingsOptions));
+            ExtraParameters? extraParams = null;
+
+            // CUSTOM CODE NOTE:
+            //   If AdditionalProperties are provided, the decision has been made to default extraParams to "PassThrough"
+            if (embeddingsOptions.AdditionalProperties != null && embeddingsOptions.AdditionalProperties.Count > 0)
+            {
+                extraParams ??= ExtraParameters.PassThrough;
+            }
 
             using RequestContent content = embeddingsOptions.ToRequestContent();
             RequestContext context = FromCancellationToken(cancellationToken);
             Response response = Embed(content, extraParams?.ToString(), context);
             return Response.FromValue(EmbeddingsResult.FromResponse(response), response);
+        }
+
+        /// <summary>
+        /// [Protocol Method] Return the embedding vectors for given text prompts.
+        /// The method makes a REST API call to the `/embeddings` route on the given endpoint.
+        /// <list type="bullet">
+        /// <item>
+        /// <description>
+        /// This <see href="https://github.com/Azure/azure-sdk-for-net/blob/main/sdk/core/Azure.Core/samples/ProtocolMethods.md">protocol method</see> allows explicit creation of the request and processing of the response for advanced scenarios.
+        /// </description>
+        /// </item>
+        /// <item>
+        /// <description>
+        /// Please try the simpler <see cref="EmbedAsync(EmbeddingsOptions,CancellationToken)"/> convenience overload with strongly typed models first.
+        /// </description>
+        /// </item>
+        /// </list>
+        /// </summary>
+        /// <param name="content"> The content to send as the body of the request. </param>
+        /// <param name="extraParams">
+        /// Controls what happens if extra parameters, undefined by the REST API,
+        /// are passed in the JSON request payload.
+        /// This sets the HTTP request header `extra-parameters`. Allowed values: "error" | "drop" | "pass-through"
+        /// </param>
+        /// <param name="context"> The request context, which can override default behaviors of the client pipeline on a per-call basis. </param>
+        /// <exception cref="ArgumentNullException"> <paramref name="content"/> is null. </exception>
+        /// <exception cref="RequestFailedException"> Service returned a non-success status code. </exception>
+        /// <returns> The response returned from the service. </returns>
+        public virtual async Task<Response> EmbedAsync(RequestContent content, string extraParams = null, RequestContext context = null)
+        {
+            Argument.AssertNotNull(content, nameof(content));
+
+            using var scope = ClientDiagnostics.CreateScope("EmbeddingsClient.Embed");
+            scope.Start();
+            try
+            {
+                using HttpMessage message = CreateEmbedRequest(content, extraParams, context);
+                return await _pipeline.ProcessMessageAsync(message, context).ConfigureAwait(false);
+            }
+            catch (Exception e)
+            {
+                scope.Failed(e);
+                throw;
+            }
+        }
+
+        /// <summary>
+        /// [Protocol Method] Return the embedding vectors for given text prompts.
+        /// The method makes a REST API call to the `/embeddings` route on the given endpoint.
+        /// <list type="bullet">
+        /// <item>
+        /// <description>
+        /// This <see href="https://github.com/Azure/azure-sdk-for-net/blob/main/sdk/core/Azure.Core/samples/ProtocolMethods.md">protocol method</see> allows explicit creation of the request and processing of the response for advanced scenarios.
+        /// </description>
+        /// </item>
+        /// <item>
+        /// <description>
+        /// Please try the simpler <see cref="Embed(EmbeddingsOptions,CancellationToken)"/> convenience overload with strongly typed models first.
+        /// </description>
+        /// </item>
+        /// </list>
+        /// </summary>
+        /// <param name="content"> The content to send as the body of the request. </param>
+        /// <param name="extraParams">
+        /// Controls what happens if extra parameters, undefined by the REST API,
+        /// are passed in the JSON request payload.
+        /// This sets the HTTP request header `extra-parameters`. Allowed values: "error" | "drop" | "pass-through"
+        /// </param>
+        /// <param name="context"> The request context, which can override default behaviors of the client pipeline on a per-call basis. </param>
+        /// <exception cref="ArgumentNullException"> <paramref name="content"/> is null. </exception>
+        /// <exception cref="RequestFailedException"> Service returned a non-success status code. </exception>
+        /// <returns> The response returned from the service. </returns>
+        public virtual Response Embed(RequestContent content, string extraParams = null, RequestContext context = null)
+        {
+            Argument.AssertNotNull(content, nameof(content));
+
+            using var scope = ClientDiagnostics.CreateScope("EmbeddingsClient.Embed");
+            scope.Start();
+            try
+            {
+                using HttpMessage message = CreateEmbedRequest(content, extraParams, context);
+                return _pipeline.ProcessMessage(message, context);
+            }
+            catch (Exception e)
+            {
+                scope.Failed(e);
+                throw;
+            }
         }
 
         private class AddApiKeyHeaderPolicy : HttpPipelinePolicy